Split off htop to a separate (EPEL-only) list for CentOS 7
On CentOS 7, htop is not part of the base distribution, and installing
htop therefore requires the EPEL repositories.

ansible-role-epel is therefore declared as a dependency now. Note that
this will not enable the EPEL repos by default; that only happens if
`epel_repos_enabled_by_default` is set to True.